requirements/requirements.txt
mysql-connector-python==9.4.0
numpy==2.3.2
pandas==2.3.1
py4j==0.10.9.7
pyspark==3.5.6
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
spotipy==2.25.1

